+++ examples/audio_player_interaction/lib/sound_player.dart
@@ -1,4 +1,5 @@
 import 'package:audioplayers/audioplayers.dart';
+import 'package:flutter/foundation.dart';
 
 abstract class SoundPlayer {
   bool get isPlaying;
@@ -10,8 +11,7 @@ abstract class SoundPlayer {
 }
 
 class AudioSoundPlayer implements SoundPlayer {
-  final AudioCache player = AudioCache();
-  AudioPlayer? _audio;
+  final AudioPlayer player = AudioPlayer();
   void Function()? _onStop;
 
   @override
@@ -25,24 +25,34 @@ class AudioSoundPlayer implements SoundPlayer {
   @override
   Future play(String assetToPlay, {bool loop = false}) async {
     if (loop) {
-      _audio = await player.loop(assetToPlay);
+      await player.play(AssetSource(assetToPlay));
     } else {
-      _audio = await player.play(
+      await player.play(AssetSource(
         assetToPlay,
-      );
+      ));
     }
-    _audio?.onPlayerStateChanged
-        .listen((s) => s == PlayerState.COMPLETED ? _onStop?.call() : null);
+    player.onPlayerStateChanged.listen((s) => _listen(s));
+  }
+
+  void _listen(PlayerState s) {
+    logIt('Player state changed: $s');
+    if (s == PlayerState.completed) {
+      stop();
+    }
   }
 
   @override
   Future stop() async {
-    _audio?.stop();
+    _onStop?.call();
   }
 
   @override
   Future setAsset(String assetPath,
       {bool preload = true, Duration? initialPosition}) async {
     // return player.setAsset(assetPath);
+  }
+
+  void logIt(String msg) {
+    debugPrint('SoundLoop: $msg');
   }
 }