+++ speech_to_text/android/src/main/kotlin/com/csdcorp/speech_to_text/SpeechToTextPlugin.kt
@@ -18,10 +18,7 @@ import android.os.Build
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
-import android.speech.RecognitionListener
-import android.speech.RecognitionService
-import android.speech.RecognizerIntent
-import android.speech.SpeechRecognizer
+import android.speech.*
 import android.speech.SpeechRecognizer.createOnDeviceSpeechRecognizer
 import android.speech.SpeechRecognizer.createSpeechRecognizer
 import android.util.Log
@@ -37,10 +34,10 @@ import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 import io.flutter.plugin.common.PluginRegistry
-import io.flutter.plugin.common.PluginRegistry.Registrar
 import org.json.JSONArray
 import org.json.JSONObject
 import java.util.*
+import java.util.concurrent.Executors
 
 
 enum class SpeechToTextErrors {
@@ -127,25 +124,6 @@ public class SpeechToTextPlugin :
         onAttachedToEngine(flutterPluginBinding.getApplicationContext(), flutterPluginBinding.getBinaryMessenger());
     }
 
-    // This static function is optional and equivalent to onAttachedToEngine. It supports the old
-    // pre-Flutter-1.12 Android projects. You are encouraged to continue supporting
-    // plugin registration via this function while apps migrate to use the new Android APIs
-    // post-flutter-1.12 via https://flutter.dev/go/android-project-migration.
-    //
-    // It is encouraged to share logic between onAttachedToEngine and registerWith to keep
-    // them functionally equivalent. Only one of onAttachedToEngine or registerWith will be called
-    // depending on the user's project. onAttachedToEngine or registerWith must both be defined
-    // in the same class.
-    companion object {
-        @JvmStatic
-        fun registerWith(registrar: Registrar) {
-            val speechPlugin = SpeechToTextPlugin()
-            speechPlugin.currentActivity = registrar.activity()
-            registrar.addRequestPermissionsResultListener(speechPlugin)
-            speechPlugin.onAttachedToEngine(registrar.context(), registrar.messenger())
-        }
-    }
-
     private fun onAttachedToEngine(applicationContext: Context, messenger: BinaryMessenger) {
         this.pluginContext = applicationContext;
         channel = MethodChannel(messenger, pluginChannelName)
@@ -205,6 +183,7 @@ public class SpeechToTextPlugin :
                     if (null == localeId) {
                         localeId = defaultLanguageTag
                     }
+                    localeId = localeId.replace( '_', '-')
                     var partialResults = call.argument<Boolean>("partialResults")
                     if (null == partialResults) {
                         partialResults = true
@@ -288,15 +267,14 @@ public class SpeechToTextPlugin :
             result.success(false)
             return
         }
+        var listenMode = enumValues<ListenMode>()[listenModeIndex]
+
         resultSent = false
-        createRecognizer(onDevice)
+        createRecognizer(onDevice, listenMode)
         minRms = 1000.0F
         maxRms = -100.0F
         debugLog("Start listening")
-        var listenMode = ListenMode.deviceDefault
-        if ( listenModeIndex == ListenMode.dictation.ordinal) {
-            listenMode = ListenMode.dictation
-        }
+
         optionallyStartBluetooth()
         setupRecognizerIntent(languageTag, partialResults, listenMode, onDevice )
         handler.post {
@@ -371,14 +349,39 @@ public class SpeechToTextPlugin :
             result.success(false)
             return
         }
-        var detailsIntent = RecognizerIntent.getVoiceDetailsIntent(pluginContext)
-        if (null == detailsIntent) {
-            detailsIntent = Intent(RecognizerIntent.ACTION_GET_LANGUAGE_DETAILS)
-            detailsIntent.setPackage("com.google.android.googlequicksearchbox")
+        var hasPermission = ContextCompat.checkSelfPermission(pluginContext!!,
+            Manifest.permission.RECORD_AUDIO) == PackageManager.PERMISSION_GRANTED
+        if (Build.VERSION.SDK_INT >= 33 && hasPermission) {
+            if ( SpeechRecognizer.isOnDeviceRecognitionAvailable(pluginContext!!)) {
+                // after much experimentation this was the only working iteration of the
+                // checkRecognitionSupport that works.
+            var recognizer = createOnDeviceSpeechRecognizer(pluginContext!!)
+            var recognizerIntent = Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH)
+//            var recognizer = createSpeechRecognizer(pluginContext!!)
+//            var recognizerIntent = Intent(RecognizerIntent.ACTION_GET_LANGUAGE_DETAILS)
+            recognizer?.checkRecognitionSupport(recognizerIntent, Executors.newSingleThreadExecutor(),
+                object : RecognitionSupportCallback {
+                    override fun onSupportResult(recognitionSupport: RecognitionSupport) {
+                        var details = LanguageDetailsChecker( result, debugLogging )
+                        details.createResponse(recognitionSupport.supportedOnDeviceLanguages )
+                        recognizer?.destroy()
+                    }
+                    override fun onError(error: Int) {
+                        debugLog("error from checkRecognitionSupport: " + error)
+                        recognizer?.destroy()
+                    }
+                })
+            }
+        } else {
+            var detailsIntent = RecognizerIntent.getVoiceDetailsIntent(pluginContext)
+            if (null == detailsIntent) {
+                detailsIntent = Intent(RecognizerIntent.ACTION_GET_LANGUAGE_DETAILS)
+                detailsIntent.setPackage("com.google.android.googlequicksearchbox")
+            }
+            pluginContext?.sendOrderedBroadcast(
+                    detailsIntent, null, LanguageDetailsChecker(result, debugLogging),
+                    null, Activity.RESULT_OK, null, null)
         }
-        pluginContext?.sendOrderedBroadcast(
-                detailsIntent, null, LanguageDetailsChecker(result, debugLogging),
-                null, Activity.RESULT_OK, null, null)
     }
 
     private fun notifyListening(isRecording: Boolean ) {
@@ -566,7 +569,7 @@ public class SpeechToTextPlugin :
         return list.firstOrNull()?.serviceInfo?.let { ComponentName(it.packageName, it.name) }
     }
 
-    private fun createRecognizer(onDevice: Boolean) {
+    private fun createRecognizer(onDevice: Boolean, listenMode: ListenMode) {
         if ( null != speechRecognizer && onDevice == lastOnDevice ) {
             return
         }
@@ -613,7 +616,7 @@ public class SpeechToTextPlugin :
             }
         }
         debugLog("before setup intent")
-        setupRecognizerIntent(defaultLanguageTag, true, ListenMode.deviceDefault, false )
+        setupRecognizerIntent(defaultLanguageTag, true, listenMode, false )
         debugLog("after setup intent")
     }
 
@@ -629,7 +632,12 @@ public class SpeechToTextPlugin :
                 run {
                     recognizerIntent = Intent(RecognizerIntent.ACTION_RECOGNIZE_SPEECH).apply {
                         debugLog("In RecognizerIntent apply")
-                        putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL, RecognizerIntent.LANGUAGE_MODEL_FREE_FORM)
+                        if (listenMode == ListenMode.search) {
+                            putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL, RecognizerIntent.LANGUAGE_MODEL_WEB_SEARCH)
+                        }
+                        else {
+                            putExtra(RecognizerIntent.EXTRA_LANGUAGE_MODEL, RecognizerIntent.LANGUAGE_MODEL_FREE_FORM)
+                        }
                         debugLog("put model")
                         val localContext = pluginContext
                         if (null != localContext) {
@@ -778,7 +786,7 @@ class LanguageDetailsChecker(flutterResult: Result, logging: Boolean ) : Broadca
         }
     }
 
-    private fun createResponse(supportedLanguages: List<String>?) {
+    public fun createResponse(supportedLanguages: List<String>?) {
         val currentLocale = Locale.getDefault()
         val localeNames = ArrayList<String>()
         localeNames.add(buildIdNameForLocale(currentLocale))