+++ speech_to_text/test/test_speech_to_text_platform.dart
@@ -37,9 +37,9 @@ class TestSpeechToTextPlatform extends SpeechToTextPlatform {
   static const String finalRecognizedJson =
       '{"alternates":[{"recognizedWords":"$secondRecognizedWords","confidence":$secondConfidence}],"finalResult":true}';
   static const SpeechRecognitionWords firstWords =
-      SpeechRecognitionWords(firstRecognizedWords, firstConfidence);
+      SpeechRecognitionWords(firstRecognizedWords, null, firstConfidence);
   static const SpeechRecognitionWords secondWords =
-      SpeechRecognitionWords(secondRecognizedWords, secondConfidence);
+      SpeechRecognitionWords(secondRecognizedWords, null, secondConfidence);
   static final SpeechRecognitionResult firstRecognizedResult =
       SpeechRecognitionResult([firstWords], false);
   static final SpeechRecognitionResult secondRecognizedResult =
@@ -83,7 +83,8 @@ class TestSpeechToTextPlatform extends SpeechToTextPlatform {
       partialResults = true,
       onDevice = false,
       int listenMode = 0,
-      sampleRate = 0}) async {
+      sampleRate = 0,
+      SpeechListenOptions? options}) async {
     listenInvoked = true;
     listenLocale = localeId;
     if (listenException) {